+++ src/main/java/net/mindsoup/dndata/services/ClaimService.java
@@ -6,8 +6,8 @@ import net.mindsoup.dndata.models.dao.User;
  * Created by Valentijn on 11-8-2019
  */
 public interface ClaimService {
-	User getClaimOn(Long objectId);
-	boolean claim(Long objectId, User user);
+	User getClaimOn(Integer objectId);
+	boolean claim(Integer objectId, User user);
 	void clearClaimForUser(Long userId);
 	void clearOldClaims();
 }
